(chore): fix reconcile plugin on plugin definition change (#1406)
(chore): update fixtures

(chore): split plugin e2e into sub scenarios

(chore): use scenarios in plugin e2e

(chore): remove unused import

(chore): add license headers

(chore): fix lint errors

(chore): fix lint errors

(chore): fix typo

(chore): simplify watches

(chore): fix tests

(chore): tidy up!

(chore): fix flux scenario e2e test

(chore): fix flux scenario e2e test

(chore): debug helm release in actions

(chore): wait for flux helm controller deployment to be available

(chore): generate flux pod logs

(chore): use label constants

(chore): print pod logs on test failure

(chore): print pod logs on test failure

(chore): adds failure check on print pod logs

(chore): merge helm and flux controllers

(chore): unique test resource names

(chore): fix multiple scheme writes

(chore): fix lint errors

(chore): remove multiple scheme registration

(chore): remove err context

(chore): remove label selector in enqueue request

(chore): keep concurrent reconcile to 10

(chore): remove wait for flux deployment

(chore): adds github token env in e2e workflow

(chore): add echo stmt

(chore): fix line termination

(chore): add github token env to unit tests

(cherry picked from commit a4fc9c1b7a2224ee3511126284d9992dc845a4a1)
