+++ .github/workflows/unit-tests.yml
@@ -27,6 +27,8 @@ jobs:
 
   test:
     runs-on: [ default ]
+    env:
+      GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
     steps:
       - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4
         with:
