+++ .golangci.yaml
@@ -103,6 +103,7 @@ linters:
       dot-import-whitelist:
         - github.com/onsi/ginkgo/v2
         - github.com/onsi/gomega
+        - github.com/onsi/gomega/gstruct
     usestdlibvars:
       http-method: true
       http-status-code: true
